Tests refactoring, EthApi refactoring.
